requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pillow==12.1.0
httpx==0.28.1
python-multipart==0.0.21